Restructure according to package diagram, with respect to system architecture
